moved all config to a single folder
