+++ duqtools/config/__init__.py
@@ -1,0 +1,3 @@
+from .config import Config, WorkDirectory, cfg
+
+__all__ = ['cfg', 'WorkDirectory', 'Config']
