+++ duqtools/config/basemodel.py
