+++ duqtools/config/status.py
@@ -1,0 +1,9 @@
+from .basemodel import BaseModel
+
+
+class StatusConfig(BaseModel):
+    """Status_config."""
+
+    msg_completed: str = 'Status : Completed successfully'
+    msg_failed: str = 'Status : Failed'
+    msg_running: str = 'Status : Running'
